fix: cambiado tipo unsigned long a int

+++ src/ultrasonico/sensor_ultrasonico.h
@@ -1,9 +1,9 @@
 #ifndef SENSOR_ULTRASONICO_H
 #define SENSOR_ULTRASONICO_H
 
-void inicializarUltrasonico(int trig, unsigned long echo);
-void estadoInicial(int trig, unsigned long echo);
+void inicializarUltrasonico(int trig, int echo);
+void estadoInicial(int trig, int echo);
 void enviarPulso(int trig);
-float calcularDistancia(unsigned long echo);
+float calcularDistancia(int echo);
 
 #endif
